Renamed CDP event class. Remove dead code
